move body into monitor
